Move visitor interface to a separate file
